requirements.txt
ttkbootstrap
pyperclip
pillow
